create helper file with default configuration
