+++ web-app/server-routes/maindata-router.js
@@ -3,38 +3,17 @@ const { v4: uuidv4 } = require("uuid");
 const mysqlx = require("@mysql/xdevapi");
 const dns = require("dns").promises;
 const memcachePlus = require("memcache-plus");
+const {
+  dbConfig,
+  dbSessionConfig,
+  executeQuery,
+  executeSimpleQuery,
+  cacheDefaultTTL,
+  memcachedConfig,
+} = require("./helper");
 
 // Erstellen einer Express Router Instanz
 const router = express.Router();
-
-// ------------------------------------------------------------
-// Datenbank Initialisierung
-// ------------------------------------------------------------
-
-const dbConfig = {
-  host: "my-app-mysql-service",
-  port: "33060",
-  user: "root",
-  password: "mysecretpw",
-  schema: "popular",
-};
-
-const dbSessionConfig = {
-  host: "my-app-mysql-service",
-  port: "33060",
-  user: "root",
-  password: "mysecretpw",
-};
-
-async function executeQuery(query, data) {
-  let session = await mysqlx.getSession(dbConfig);
-  return await session.sql(query, data).bind(data).execute();
-}
-
-async function executeSimpleQuery(query) {
-  let session = await mysqlx.getSession(dbConfig);
-  return await session.sql(query).bind(data).execute();
-}
 
 // -------------------------------------------------------
 // Memcache Configuration
@@ -43,13 +22,6 @@ async function executeSimpleQuery(query) {
 //Connect to the memcached instances
 let memcached = null;
 let memcachedServers = [];
-const cacheDefaultTTL = 15;
-
-const memcachedConfig = {
-  host: "my-memcached-service",
-  port: 11211,
-  updateInterval: 5000,
-};
 
 async function getMemcachedServersFromDns() {
   try {
